Changes file structure, renames things, & refactors to use logger
