+++ src/api/controllers/quiz.controller.js
@@ -1,10 +1,12 @@
 const quizService = require('../services/quiz.service')
+const logger = require('../../shared/logger')
+
 
 const getAll = async (req, res, next) => {
     try{
         res.json(await quizService.retrieveAll())
     } catch (err){
-        console.error('Error in quiz.controller getAll: ', err.message)
+        logger.error('Error in quiz.controller getAll: ', err.message)
         next(err)
     }
 }
@@ -13,7 +15,7 @@ const getById = async (req, res, next) => {
   try{
       res.json(await quizService.retrieveByID(req.params.id))
   } catch(err){
-      console.error('Error in quiz.controller.getById:', err.message)
+      logger.error('Error in quiz.controller.getById:', err.message)
       next(err)
   }
 }
@@ -22,7 +24,7 @@ const post = async (req, res, next) => {
     try{
         res.json(await quizService.create(req.body))
     } catch(err){
-        console.error('Error in quiz.controller.post:', err.message)
+        logger.error('Error in quiz.controller.post:', err.message)
         next(err)
     }
 }
@@ -32,7 +34,7 @@ const remove = async (req, res, next) => {
     try{
         res.json(await quizService.remove(req.params.id))
     } catch(err){
-        console.error('Error in quiz.controller.remove:', err.message)
+        logger.error('Error in quiz.controller.remove:', err.message)
         next(err)
     }
 }
@@ -41,7 +43,7 @@ const put = async (req, res, next) => {
     try{
         res.json(await quizService.update(req.params.id, req.body))
     } catch(err){
-        console.error('Error in quiz.controller.put:', err.message)
+        logger.error('Error in quiz.controller.put:', err.message)
         next(err)
     }
 }
